Listening for SE gives a time window before maintenance
Clarifying that listening for schedule events gives callers an extended timeout before maintenance is performed as this was not clear previously. Also the guidance around having to periodically query the endpoint for changes was missing. Finally highlighted that the DocumentIncarnation tag provides an easy way to detect if the events payload has changed.